+++ typeguard/__init__.py
@@ -312,7 +312,7 @@ def check_typed_dict(argname: str, value, expected_type, memo: Optional[_CallMem
             keys_formatted = ', '.join('"{}"'.format(key) for key in sorted(missing_keys))
             raise TypeError('required key(s) ({}) missing from {}'.format(keys_formatted, argname))
 
-    for key, argtype in expected_type.__annotations__.items():
+    for key, argtype in get_type_hints(expected_type).items():
         argvalue = value.get(key, _missing)
         if argvalue is not _missing:
             check_type('dict item "{}" for {}'.format(key, argname), argvalue, argtype)