+++ typeguard/__init__.py
@@ -592,9 +592,6 @@ def check_type(argname: str, value, expected_type, memo: Optional[_TypeCheckMemo
     elif isclass(expected_type):
         if issubclass(expected_type, Tuple):
             check_tuple(argname, value, expected_type, memo)
-        elif issubclass(expected_type, Callable) and hasattr(expected_type, '__args__'):
-            # Needed on Python 3.5.0 to 3.5.2
-            check_callable(argname, value, expected_type, memo)
         elif issubclass(expected_type, (float, complex)):
             check_number(argname, value, expected_type)
         elif _subclass_check_unions and issubclass(expected_type, Union):