+++ docs/versionhistory.rst
@@ -6,6 +6,7 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
 **UNRELEASED**
 
 - Added support for ``typing.NoReturn``
+- Fixed forward references in ``TypedDict`` not being resolved
 
 **2.8.0** (2020-06-02)