+++ docs/versionhistory.rst
@@ -9,6 +9,7 @@ This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-v
 - Fixed erroneous ``TypeError`` when trying to check against non-runtime ``typing.Protocol``
   (skips the check for now until a proper compatibility check has been implemented)
 - Fixed forward references in ``TypedDict`` not being resolved
+- Fixed checking against recursive types
 
 **2.8.0** (2020-06-02)