+++ docs/versionhistory.rst
@@ -3,7 +3,7 @@ Version history
 
 This library adheres to `Semantic Versioning 2.0 <https://semver.org/#semantic-versioning-200>`_.
 
-**UNRELEASED**
+**2.9.0** (2020-06-06)
 
 - Upped the minimum Python version from 3.5.2 to 3.5.3
 - Added support for ``typing.NoReturn``